programs/operator-dashboard/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "operator-dashboard"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "oracle-shared",
    "redis[hiredis]>=5.0",
    "python-dotenv>=1.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "python-telegram-bot>=21.0",
]

[tool.hatch.build.targets.wheel]
packages = ["operator_dashboard"]
